Move action helpers into their own module
